requirements.txt
Flask
PyPDF2
python-docx
spacy
language-tool-python
Werkzeug
pdfkit
